+++ mystack.h
@@ -20,7 +20,7 @@ This file is part of Reverse Notation Calc.
 #define _mystack_h
 
 typedef struct {
-    int *ptr;
+    double *ptr;
     size_t size;
     size_t count;
 } my_stack;
@@ -28,10 +28,10 @@ typedef struct {
 void my_stack_init(my_stack* stack, size_t size);
 void my_stack_done(my_stack* stack);
 
-void my_stack_push(my_stack* stack, int value);
-int my_stack_pull(my_stack* stack);
+void my_stack_push(my_stack* stack, double value);
+double my_stack_pull(my_stack* stack);
 
-int my_stack_top(my_stack* stack);
+double my_stack_top(my_stack* stack);
 int my_stack_is_empty(my_stack* stack);
 int my_stack_is_full(my_stack* stack);
